refactor toward composition
