+++ training_templates/data_utils.py
@@ -1,4 +1,5 @@
 import os
+from typing import Tuple
 
 import pandas as pd
 from pyspark.sql import SparkSession
@@ -9,9 +10,10 @@ from pyspark.sql.types import (
     StructField,
     StructType,
 )
+from sklearn.model_selection import train_test_split
 
 
-def train_test_split(
+def spark_train_test_split(
     feature_table_name: str,
     unique_id: str,
     train_val_size: float,
@@ -66,6 +68,26 @@ def train_test_split(
     )
 
     return (train_output_table_name, test_output_table_name)
+
+
+def train_val_split(pandas_df: pd.DataFrame, label_col: str, train_size: float, 
+                    shuffle: int, random_state: int) -> Tuple[pd.DataFrame, pd.DataFrame, pd.Series, pd.Series]:
+    """
+    Split a Pandas DataFrame of features into training and validation datasets
+    """
+
+    non_label_cols = [
+        col for col in pandas_df.columns if col != label_col
+    ]
+    X_train, X_val, y_train, y_val = train_test_split(
+        pandas_df[non_label_cols],
+        pandas_df[label_col],
+        train_size=train_size,
+        random_state=random_state,
+        shuffle=shuffle,
+    )
+
+    return (X_train, X_val, y_train, y_val)
