+++ training_templates/metrics.py
@@ -2,14 +2,14 @@ from collections import OrderedDict
 from sklearn.metrics import precision_recall_fscore_support
 
 
-def xgboost_classification_metrics(model, X_transformed, y, average="weighted", round_digits=3):
+def xgboost_classification_metrics(model, x, y, average="weighted", round_digits=3):
 
     best_iteration = model.best_iteration
     best_xgboost_rounds = (0, best_iteration + 1)
      
     precision_val, recall_val, f1_val, _ = precision_recall_fscore_support(
                     y,
-                    model.predict(X_transformed, iteration_range=best_xgboost_rounds),
+                    model.predict(x, iteration_range=best_xgboost_rounds),
                     average=average,
                 )
 
@@ -22,12 +22,12 @@ def xgboost_classification_metrics(model, X_transformed, y, average="weighted", 
     return metrics
 
 
-def sklearn_classification_metrics(model, X_transformed, y, average="weighted", round_digits=3):
+def sklearn_classification_metrics(model, x, y, average="weighted", round_digits=3):
      
     precision_val, recall_val, f1_val, _ = precision_recall_fscore_support(
                     y, 
-                    model.predict(X_transformed), 
-                    average="weighted"
+                    model.predict(x), 
+                    average=average
             )
 
     metrics = OrderedDict()
